feat(api)!: move to generic refund status API
The refund batch status API is at path `GET /refund/batch/{batchID}`,
which while continuing to work, will no longer return `batchID` as
part of the response. This allows the internal API to be more generic,
supporting fetching the refund status by more identifiers.

The path for the internal API has now changed to:
`GET /refund/{identifierType}/{identifierValue}`

The allowed values for `identifierType` at this point are:
- batch
- bill

If you were not using the `batchID` in the response in your code,
then your integration will continue to work. In either case, the
`getRefundBatchStatus` is deprecated and it is recommended to use
the new `getRefundStatusByIdentifier` method instead.